Creating a file for OptionParser and helper functions
